updates

Related work items: #61

+++ app/app/src/main/java/com/kotlin101/group2/grocerylist/data/api/models/UpdateItemRequest.kt
@@ -5,6 +5,6 @@ data class UpdateItemRequest(
     val img: Any,
     val imgURL: Any,
     val name: String,
-    val pricePerUnit: Int,
+    val pricePerUnit: Double,
     val quantity: Int
 )
